Moved everything back to main and implemented the error handling according to the std::process:Termination spec
